backend/requirements.txt
Flask
APIFlask
SQLAlchemy
Flask-SQLAlchemy
flask-cors
elasticsearch
whoosh-reloaded
flask-jwt-extended
Flask-Migrate